1335: Fixing Test Trusted Directory to be compatible with IG 2024.3
Removing the bcprov-jdk18on-1.74.jar from the IG Docker image as this clashes with the version used by SAPI-G. This dependency was previously not present in the IG docker image.

Going forwards, when we upgrade IG we will need to update th rm command to delete the correct version.

Fixing the Test Trusted Directory JWKS route to set the key usage as VERIFY. The key usage of SIGN is used to expose a private key in a JWKS, which we do not want to do, we only wish to expose the public key portion which clients can use to verify sigs.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1335