1393: Using URI instead of URL for jwks_uri and redirect_uri values
URIs are a better choice as the equals method compares the string representation.

URLs on the other hand do DNS lookups to resolve the IP address of the host, and use this value in the equality check, this adds both a performance hit and causes URLs with the same representation not to be equal if the IP address changes (which can be a common scenario when DNS round robin is being used).

This issue with URL equality was causing us to suffer cache misses for the JWKS caching functionality, and caused duplicate entries to be cached.

Changing to use URIs across the codebase for consistency.

Note, URLs were originally chosen to match the public API of the library class: org.forgerock.json.jose.jwk.JWKSetParser

https://github.com/SecureApiGateway/SecureApiGateway/issues/1393