1305: The client's MTLS certificate is now validated on the /par route.
Refactored the TokenEndpointTransportCertValidationFilter, renamed it to ResponsePathTransportCertValidationFilter and reused it for both the token endpoint and par routes.

The ResponsePathTransportCertValidationFilter expects the ApiClient to be in the attributes context, as fetching the ApiClient is the only route specific piece of logic in the filter.

Installed the ResponsePathTransportCertValidationFilter in the PAR route to enforce MTLS cert validation.

Created new filter: TokenEndpointResponseFetchApiClientFilter which is responsible for fetching the ApiClient. This filter needs to be installed after the validator filter, so that it executes first on the response path and can properly install the ApiClient.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1305