1402: Adding test RegistrationRequestBuilderFilterTest that demonstrates the race condition in the filter
Test case: shouldSucceedWhenCalledConcurrently , demonstrates that the filter has a race condition. The filter is invoked by multiple threads, each task submits a unique registration request payload and expects to the RegistrationRequest object created by the filter to match this data.

The test case has a numThreads parameter which is used to demonstrate that the test works in when a single thread is used.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1402