1305: Apply mTLS check for PAR only when mTLS is the auth method
The token endpoint needs to always enforce mTLS checks, as this is the mechanism used to sender constrain access tokens. Therefore, even when private_key_jwt is used as the auth method, the mTLS cert still needs to be supplied.

The PAR endpoint differs slightly to the token endpoint because there is nothing that needs sender constraining. Therefore, mTLS checks only need to be applied if the client is using tls_client_auth as the auth method.

This change allows the ResponsePathTransportCertValidationFilter, which is used by both endpoints, to be configured to indicate whether the mTLS certificate is mandatory.

New endpoint specific heaplets have been created which can be used in routes to create the following filters: TokenEndpointTransportCertValidationFilter and ParEndpointTransportCertValidationFilter.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1305