1512: Make pass-through route send all traffic to Identity Platform
Rename 29-as-passthrough.json to 99-platform-passthrough.json

The route is now responsible for passing any requests, which have not already been routed, through to the Identity Platform.

This includes AM endpoints such as the XUI and IDM endpoints to fetch config used to render the XUI.

Fix name of 98-kube-probe.json route to ensure that it is ordered before
the pass through route.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1512